+++ Editor/Window/AnywherePage.cs
@@ -4,7 +4,7 @@
 using AmazonGameLift.Editor;
 using UnityEngine.UIElements;
 
-namespace Editor.Resources.EditorWindow.Pages
+namespace Editor.Window
 {
     public class AnywherePage
     {
