+++ Editor/Window/ConnectToFleetInput.cs
@@ -4,7 +4,7 @@ using Amazon.GameLift.Model;
 using AmazonGameLift.Editor;
 using UnityEngine.UIElements;
 
-namespace Editor.Resources.EditorWindow.Pages
+namespace Editor.Window
 {
     public class ConnectToFleetInput : StatefulInput
     {
