+++ Editor/Window/GameLiftRequestAdapter.cs.meta
