+++ Editor/Window/LandingPage.cs
@@ -5,7 +5,7 @@ using AmazonGameLift.Editor;
 using UnityEngine;
 using UnityEngine.UIElements;
 
-namespace Editor.Resources.EditorWindow.Pages
+namespace Editor.Window
 {
     public class LandingPage
     {
